Add ability to increase opacity of square
Added incrementOpacity function which increases opacity of a square by 10% every time it is colored.
Square listener now calls this function when user colors a square.

Removed border from squares and added it to the etch-a-sketch container.

Made the body surrounding the etch-a-sketch green and changed the etch-a-sketch container's overflow
to hidden so the corner pieces don't 'break' the border.